optimized city screen for smart phone (#12315)
* optimized city screen for smart phone
- moved "buy" button to city info table
- removed "add to queue" button
- expand icon changed to android defaults
- CityStatsTable: big scrollable area, expandable

* made CityStatsTable collapsible

* Extracted BuyButtonFactory and re-added buy button close to construction queue

---------
